+++ library/laboratory/src/main/java/io/mehow/laboratory/InMemoryFeatureStorage.kt
@@ -14,7 +14,7 @@ internal class InMemoryFeatureStorage : FeatureStorage {
 
   override suspend fun <T : Feature<*>> getFeatureName(feature: Class<T>) = features[feature]
 
-  override suspend fun <T : Feature<*>> setFeatures(vararg options: T): Boolean {
+  override suspend fun <T : Feature<*>> setOptions(vararg options: T): Boolean {
     for (feature in options) {
       this.features += feature.javaClass to feature.name
     }